Move keybinds directory into renderer/src
