+++ src/renderer/src/keybinds/constants.ts
